stock_analyst_agent/requirements.txt
langgraph
langchain
langchain_openai
langchain-community
langchain-google-genai
ollama
pandas
datetime
ipython
python-dotenv
matplotlib
yfinance
ta